Extract FallbackStep
